requirements.txt
click==8.1.2
decorator==5.1.1
grpcio==1.44.0
grpcio-reflection==1.44.0
grpcio-tools==1.44.0
pika==1.2.0
plumbum==1.7.2
protobuf==3.20.1
py==1.11.0
pyclean==2.2.0
retry==0.9.2
six==1.16.0
typer==0.4.1
